AXON/requirements.txt
fastapi
uvicorn
Pillow
torch
opencv-python-headless
ultralytics
